Make settings respect param server, lock in serial
Settings behavior has changed to:
1) initially query device for settings
2) make settings match those explicitly set in param server
3) update param server with newly merged settings

After initial startup, parameter server should have all settings in it. This makes subsequent disconnect/reconnect cycles have the exact same settings that the device had before the disconnect/reconnect cycle, even if device settings changed or were lost.

Also, once the driver is connected to a device, it will only connect to that exact device on a subsequent disconnect/reconnect cycle. If you need to swap devices, you need to restart the driver. If the parameter server is not restarted as well and the node name doesn't change, then the old settings will be sent to the new device.